% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afinn}
\alias{afinn}
\title{AFINN Lexicon}
\format{
A \code{\link[base]{data.frame}} with 3,380 rows and 2 columns
}
\source{
\link{http://www2.imm.dtu.dk/pubdb/views/publication_details.php?id=6010}
}
\description{
The preprocessed and deduplicated AFINN lexicon (Nielsen 2011). The original dataset was retrieved from \pkg{syuzhet} and preprocessed using \pkg{WoCSentiment} function \code{cleanse_lex}.
}
\details{
\itemize{
\item word. An n-gram representing a term or phrase
\item value. The sentiment intensity score

}
}
\references{
Nielsen, F.Å. A new ANEW: Evaluation of a word list for sentiment analysis in microblogs, arXiv preprint arXiv:1103.2903, (2011).
}
