% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bing}
\alias{bing}
\title{Opinion Lexicon}
\format{
A \code{\link[base]{data.frame}} with 6,782 rows and 2 columns
}
\source{
http://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html#lexicon
}
\description{
The preprocessed and deduplicated Opinion lexicon (Hu & Liu 2004). The original dataset was retrieved from \pkg{syuzhet} and preprocessed using \pkg{WoCSentiment} function \code{cleanse_lex}.
}
\details{
\itemize{
\item word. An n-gram representing a term or phrase
\item value. The sentiment polarity score (-1, 1)

}
}
\references{
Liu, B., Hu, M., and Cheng, J. "Opinion Observer: Analyzing and Comparing Opinions on the Web." Proceedings of the 14th International World Wide Web conference (WWW-2005), May 10-14, 2005, Chiba, Japan.
 
Hu, M. and Liu, B. "Mining and Summarizing Customer Reviews." Proceedings of the ACM SIGKDD International Conference on Knowledge Discovery and Data Mining (KDD-2004), Aug 22-25, 2004, Seattle, Washington, USA.
}
