% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOC_Functions_FINAL.R
\name{cleanse_lex}
\alias{cleanse_lex}
\title{Preprocess Sentiment Lexicons}
\usage{
cleanse_lex(lexicon, emot = lexicon::hash_emoticons)
}
\arguments{
\item{lexicon}{A character string identifying the lexicon to cleanse. There must be a \code{\link[base]{data.frame}} in your global environment of the same name containing two columns, "word" and "value".}

\item{emot}{A \code{\link[base]{data.frame}} containing two columns, the first column containing emoticons and the second column containing the meanings of the emoticons in words. Defaults to \pkg{lexicon} \code{\link[lexicon]{hash_emoticons}}.}
}
\value{
A \code{\link[base]{data.frame}} with "word" and "value" columns that combines duplicates and removes neutrals.
}
\description{
Cleanse lexicons of duplicate terms, neutral sentiment terms, and punctuation, where appropriate.
}
\examples{
\dontrun{
library(syuzhet)
syuzhet <- syuzhet::get_sentiment_dictionary()
syuzhet <- cleanse_lex(syuzhet)
}

}
