% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOC_Functions_FINAL.R
\name{get_all_sents}
\alias{get_all_sents}
\title{Individual Lexicon Sentiment Analysis: Dataset Scores}
\usage{
get_all_sents(dat, lexicon, lex, scoring = "abs")
}
\arguments{
\item{dat}{A \code{\link[base]{data.frame}} with column named "text" to perform sentiment analysis on}

\item{lexicon}{A text string of the lexicon name, ie "syuzhet"}

\item{lex}{A lexicon \code{\link[base]{data.frame}} containing two columns, "word" and "value"}

\item{scoring}{A character string indicating the type of scoring function to use (absolute = "abs", relative = "rel", subtraction = "sub"). Defaults to \code{scoring = "abs"}}
}
\value{
A \code{\link[base]{vector}} containing all sentiment scores for text documents in a dataset for the provided lexicon.
}
\description{
Compute Sentiment Analysis scores for all text documents in a dataset based on a chosen individual lexicon and scoring method.
}
