% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOC_Functions_FINAL.R
\name{get_WoC_lexicon}
\alias{get_WoC_lexicon}
\title{WoC Sentiment Lexicons}
\usage{
get_WoC_lexicon(lexicon)
}
\arguments{
\item{lexicon}{a character string identifying the lexicon from the default WoC crowd to return. Options include: 

\itemize{
\item \code{"afinn"}: AFINN sentiment intensity lexicon (Nielsen 2011)
\item \code{"bing"}: Opinion sentiment polarity lexicon (Hu & Liu 2004)
\item \code{"inquirer"}: General Inquirer sentiment polarity lexicon (Stone et al. 1966)
\item \code{"loughran"}: Loughran and McDonald sentiment polarity lexicon (Loughran & McDonald 2016)
\item \code{"nrc"}: NRC sentiment polarity lexicon (Mohammad & Turney 2010)
\item \code{"senticnet"}: SenticNet sentiment intensity lexicon (Cambria et al. 2016)
\item \code{"sentiwn"}: SentiWordNet sentiment intensity lexicon (Baccianella et al. 2010)
\item \code{"slang"}: SlangSD sentiment intensity lexicon (Wu et al. 2016)
\item \code{"socal"}: SO-CAL sentiment intensity lexicon (Taboada et al. 2011)
\item \code{"syuzhet"}: Jockers sentiment intensity lexicon (Jockers 2017)
\item \code{"textblob"}: TextBlob sentiment intensity lexicon (Loria 2017)
\item \code{"vader"}: Vader sentiment intensity lexicon (Hutto & Gilbert 2014)
\item \code{"wkwsci"}: WKWSCI sentiment intensity lexicon (Khoo & Johnkhan 2018)
}}
}
\value{
A \code{\link[base]{data.frame}} with:
\itemize{
\item word. n-grams representing a term or phrase
\item value. Sentiment intensity or polarity scores

}
}
\description{
Obtain an individual lexicon from the WoC lexicon crowd.
}
\examples{
get_WoC_lexicon("afinn")

}
\references{
Baccianella S., Esuli, A. and Sebastiani, F. (2010). SentiWordNet 3.0: An Enhanced Lexical Resource for Sentiment Analysis and Opinion Mining. International Conference on Language Resources and Evaluation.

Cambria, E., Poria, S., Bajpai, R. and Schuller, B. SenticNet 4: A semantic resource for sentiment analysis based on conceptual primitives. In: COLING, pp. 2666-2677, Osaka (2016)

Hu, M. and Liu, B. "Mining and Summarizing Customer Reviews." Proceedings of the ACM SIGKDD International Conference on Knowledge Discovery and Data Mining (KDD-2004), Aug 22-25, 2004, Seattle, Washington, USA.

Hutto, C.J. and Gilbert, E.E. (2014). VADER: A Parsimonious Rule-based Model
for Sentiment Analysis of Social Media Text. Eighth International Conference
on Weblogs and Social Media (ICWSM-14). Ann Arbor, MI, June 2014.

Jockers, M. L. (2017). Syuzhet: Extract sentiment and plot arcs from Text.

Khoo, C. S., & Johnkhan, S. B. (2017). Lexicon-based sentiment analysis: Comparative evaluation of six sentiment lexicons. Journal of Information Science, pp. 1–21.

Liu, B., Hu, M., and Cheng, J. "Opinion Observer: Analyzing and Comparing Opinions on the Web." Proceedings of the 14th International World Wide Web conference (WWW-2005), May 10-14, 2005, Chiba, Japan.

Loria, S. TextBlob: Simplified Text Processing (2017).

Loughran, T. and McDonald, B. (2016). Textual analysis in accounting and finance: A survey. Journal of Accounting Research 54(4), 1187-1230.

Mohammad, S. and  Turney, P. "Emotions Evoked by Common Words and Phrases: Using Mechanical Turk to Create an Emotion Lexicon." In Proceedings of the NAACL-HLT 2010 Workshop on Computational Approaches to Analysis and Generation of Emotion in Text, June 2010, LA, California.

Nielsen, F.Å. A new ANEW: Evaluation of a word list for sentiment analysis in microblogs, arXiv preprint arXiv:1103.2903, (2011).

Stone, P.J.,  Dunphy, D.C., and Smith, M.S. The general inquirer: A computer approach to content analysis, (M.I.T. Press, Cambridge, MA, 1966).

Taboada, M., Brooke, J.,  Tofiloski, M.,  Voll, K., and Stede, M. Lexicon-Based Methods for Sentiment Analysis, Computational Linguistics, 37(2) (2011) 267-307. 

Wu, L., Morstatter, F., and Liu, H. (2016). SlangSD: Building and using a sentiment dictionary of slang words for short-text sentiment classification. CoRR. abs/1168.1058. 1-15.
}
