% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOC_Functions_FINAL.R
\name{id_ngram_loc}
\alias{id_ngram_loc}
\title{Identify n-Gram Location in Text}
\usage{
id_ngram_loc(toks, sent_toks)
}
\arguments{
\item{toks}{A list of n-gram vectors, with length equal to the number of n-grams with n > 1}

\item{sent_toks}{A vector of the unigrams in the text document}
}
\value{
The starting location(s) of n-grams in the text document for the purpose of identifying the preceding token for negation identification.
}
\description{
A helper function to find the starting location(s) of n-grams in text document for the purpose of switch negation for n-grams with n > 1.
}
