% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inquirer}
\alias{inquirer}
\title{General Inquirer Lexicon}
\format{
A \code{\link[base]{data.frame}} with 3,610 rows and 2 columns
}
\source{
\link{https://inquirer.sites.fas.harvard.edu/homecat.htm}
}
\description{
The preprocessed and deduplicated General Inquirer lexicon (Stone et al. 1966). The original dataset was preprocessed using \pkg{WoCSentiment} function \code{cleanse_lex}.
}
\details{
\itemize{
\item word. An n-gram representing a term or phrase
\item value. The sentiment polarity score

}
}
\references{
Stone, P.J., Dunphy, D.C., Smith, M.S. The general inquirer: A computer approach to content analysis, (M.I.T. Press, Cambridge, MA, 1966).
}
