% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOC_Functions_FINAL.R
\name{lex_grams}
\alias{lex_grams}
\title{Individual Lexicon Sentiment Analysis: Document Score}
\usage{
lex_grams(x, lexicon, lex, n, scoring = "abs")
}
\arguments{
\item{x}{A text document}

\item{lexicon}{A text string of the lexicon name, ie "syuzhet"}

\item{lex}{A lexicon \code{\link[base]{data.frame}} with columns "word" and "value.}

\item{n}{A value indicating the maximum length of n-grams for lexicon}

\item{scoring}{A character string indicating the type of scoring function to use (absolute = "abs", relative = "rel", subtraction = "sub"). Defaults to \code{scoring = "abs"}}
}
\value{
A value indicating the aggregated sentiment score for a text document based on the identified lexicon and scoring method.
}
\description{
Obtain the text document-level sentiment score according to a chosen scoring method and individual lexicon.
}
