% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loughran}
\alias{loughran}
\title{Loughran and McDonald Lexicon}
\format{
A \code{\link[base]{data.frame}} with 2,702 rows and 2 columns
}
\source{
\link{https://sraf.nd.edu/textual-analysis/resources}
}
\description{
The preprocessed and deduplicated Loughran & McDonald lexicon (Loughran & McDonald 2016). The original dataset was retrieved from \pkg{lexicon} and preprocessed using \pkg{WoCSentiment} function \code{cleanse_lex}.
}
\details{
\itemize{
\item word. An n-gram representing a term or phrase
\item value. The sentiment polarity score

}
}
\section{License}{
 The original authors note the data is available for 
non-commercial, research use: "The data compilations provided on 
this website are for use by individual researchers.".  For more details see:
https://sraf.nd.edu/textual-analysis/resources/#Master%20Dictionary.
}

\section{Copyright}{
 Copyright holder University of Notre Dame
}

\references{
Loughran, T. and McDonald, B. (2016). Textual analysis in accounting and finance: A survey. Journal of Accounting Research 54(4), 1187-1230.
}
