% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOC_Functions_FINAL.R
\name{majvote}
\alias{majvote}
\title{Majority Voting Sentiment Analysis Scores}
\usage{
majvote(x)
}
\arguments{
\item{x}{a vector of individual lexicon sentiment scores}
}
\value{
The majority vote-based sentiment score for a text document based on a vector of individual lexicon sentiment scores
}
\description{
A helper function to compute majority voting-based sentiment scores. The assigned sentiment is based on the most frequently assigned sentiment (-1, 0, 1) from an ensemble of lexicons (the WoC crowd).
}
