% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOC_Functions_FINAL.R
\name{negates}
\alias{negates}
\title{Handle lexicon-specific switch negation}
\usage{
negates(grams, word_li, tot_words, negators = lexicon::hash_valence_shifters)
}
\arguments{
\item{grams}{A vector of n-grams of text string created using \pkg{WoCSentiment} \code{get_grams_ems}.}

\item{word_li}{A vector of sentiment-bearing n-grams based on lexicon matching}

\item{tot_words}{The number of unigrams in the text document, indicating document length.}

\item{negators}{A negators A vector of negation terms. Defaults to \pkg{lexicon} \code{\link[lexicon]{hash_valence_shifters}}, isolating those terms identified as negators (\code{y = 1}).}
}
\value{
A vector of signs (-1, 0, 1), of \code{length(word_li)}
}
\description{
Identifies negated n-grams and performs switch negation.
}
