% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrc}
\alias{nrc}
\title{NRC Lexicon}
\format{
A \code{\link[base]{data.frame}} with 5,474 rows and 2 columns
}
\source{
\link{http://saifmohammad.com/WebPages/lexicons.html}
}
\description{
The preprocessed and deduplicated NRC lexicon (Mohammad & Turney 2010). The original dataset was preprocessed using \pkg{WoCSentiment} function \code{cleanse_lex}.
}
\details{
\itemize{
\item word. An n-gram representing a term or phrase
\item value. The sentiment polarity score

}
}
\section{License}{
 The original authors note the data is available for 
non-commercial use: "If interested in commercial use of any of these lexicons, 
send email to Saif M. Mohammad (Senior Research Officer at NRC and creator of 
these lexicons): saif.mohammad@nrc-cnrc.gc.ca and Pierre Charron (Client 
Relationship Leader at NRC): Pierre.Charron@nrc-cnrc.gc.ca. A nominal 
one-time licensing fee may apply."
}

\section{Copyright}{
 Copyright holder University of Notre Dame.
}

\references{
Mohammad, S. and Turney, P. "Emotions Evoked by Common Words and Phrases: Using Mechanical Turk to Create an Emotion Lexicon." In Proceedings of the NAACL-HLT 2010 Workshop on Computational Approaches to Analysis and Generation of Emotion in Text, June 2010, LA, California.
}
