% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOC_Functions_FINAL.R
\name{perf}
\alias{perf}
\title{Obtain WoC Sentiment Analysis Performance}
\usage{
perf(scores, truesent)
}
\arguments{
\item{scores}{A \code{\link[base]{vector}} or \code{\link[base]{data.frame}} output of \code{scores_woc_pl} function}

\item{truesent}{\code{\link[base]{vector}} of known (binary) sentiment labels, where -1 = negative, 1 = positive.}
}
\value{
A \code{\link[base]{data.frame}} containing 3 rows ("Accuracy", "Sensitivity", and "Specificity") and \code{nrow(scores)} rows.
}
\description{
Obtain Accuracy, Sensitivity, and Specificity for individual lexicons in the WoC Crowd, Majority Voting Ensemble, and WoC Sentiment Analysis.
}
