% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senticnet}
\alias{senticnet}
\title{SenticNet Lexicon}
\format{
A \code{\link[base]{data.frame}} with 23,626 rows and 2 columns
}
\source{
\link{http://sentic.net/downloads}
}
\description{
The preprocessed and deduplicated SenticNet lexicon (Cambria et al. 2016). The original dataset was retrieved from \pkg{lexicon} and preprocessed using \pkg{WoCSentiment} function \code{cleanse_lex}.
}
\details{
\itemize{
\item word. An n-gram representing a term or phrase
\item value. The sentiment intensity score

}
}
\references{
Cambria, E., Poria, S., Bajpai, R. and Schuller, B. SenticNet 4: A semantic resource for sentiment analysis based on conceptual primitives. In: COLING, pp. 2666-2677, Osaka (2016)
}
