% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sentiwn}
\alias{sentiwn}
\title{SentiWordNet Lexicon}
\format{
A \code{\link[base]{data.frame}} with 19,920 rows and 2 columns
}
\source{
\link{https://sentiwordnet.isti.cnr.it}
}
\description{
The preprocessed and deduplicated SentiWordNet lexicon (Baccianella et al. 2010). The original dataset was preprocessed using \pkg{WoCSentiment} function \code{cleanse_lex}.
}
\details{
\itemize{
\item word. An n-gram representing a term or phrase
\item value. The sentiment intensity score

}
}
\section{License}{
 \link{https://creativecommons.org/licenses/by-sa/3.0/legalcode}
}

\references{
Baccianella S., Esuli, A. and Sebastiani, F. (2010). SentiWordNet 3.0: An Enhanced Lexical Resource for Sentiment Analysis and Opinion Mining. International Conference on Language Resources and Evaluation.
}
