% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slang}
\alias{slang}
\title{SlangSD Lexicon}
\format{
A \code{\link[base]{data.frame}} with 48,214 rows and 2 columns
}
\source{
\link{http://slangsd.com}
}
\description{
The preprocessed and deduplicated SlangSD lexicon (Wu et al. 2016). The original dataset was retrieved from \pkg{lexicon} and preprocessed using \pkg{WoCSentiment} function \code{cleanse_lex}.
}
\details{
\itemize{
\item word. An n-gram representing a term or phrase
\item value. The sentiment intensity score

}
}
\section{License}{
 The dictionary is free to use. If you use it for an academic publication, we 
ask that you cite it using the citation below. If it is used in anything other 
than an academic publication, we ask that you provide a credit and link to SlangSD.com.
}

\references{
Wu, L., Morstatter, F., and Liu, H. (2016). SlangSD: Building and using a sentiment dictionary of slang words for short-text sentiment classification. CoRR. abs/1168.1058. 1-15.
}
