% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{socal}
\alias{socal}
\title{SO-CAL Lexicon}
\format{
A \code{\link[base]{data.frame}} with 5,931 rows and 2 columns
}
\source{
\link{https://github.com/sfu-discourse-lab/SO-CAL}
}
\description{
The preprocessed and deduplicated SO-CAL lexicon (Taboada et al. 2011). The original dataset was preprocessed using \pkg{WoCSentiment} function \code{cleanse_lex}.
}
\details{
\itemize{
\item word. An n-gram representing a term or phrase
\item value. The sentiment intensity score

}
}
\section{License}{
 This work is licensed under a Creative Commons 
Attribution-NonCommercial-ShareAlike 4.0 International License.
https://creativecommons.org/licenses/by-nc-sa/4.0/
}

\references{
Taboada, M., Brooke, J., Tofiloski, M., Voll, K., and Stede, M. Lexicon-Based Methods for Sentiment Analysis, Computational Linguistics, 37(2) (2011) 267-307.
}
