% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{syuzhet}
\alias{syuzhet}
\title{Jockers Lexicon}
\format{
A \code{\link[base]{data.frame}} with 10,747 rows and 2 columns
}
\source{
\link{https://github.com/mjockers/syuzhet}
}
\description{
The preprocessed and deduplicated Jockers lexicon (Jockers 2017). The original \pkg{syuzhet} dataset was retrieved and preprocessed using \pkg{WoCSentiment} function \code{cleanse_lex}.
}
\details{
\itemize{
\item word. An n-gram representing a term or phrase
\item value. The sentiment intensity score

}
}
\references{
Jockers, M. L. (2017). Syuzhet: Extract sentiment and plot arcs from Text.
}
