% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOC_Functions_FINAL.R
\name{twit_pre}
\alias{twit_pre}
\title{Social Media Text Preprocessing}
\usage{
twit_pre(x)
}
\arguments{
\item{x}{A text string.}
}
\value{
A text string with mentions, retweets, hashtags, and urls removed.
}
\description{
Remove social-media specific text from a dataset prior to performing WoC Sentiment Analysis, including: mentions, retweets, hashtags, and links.
}
\examples{
text_str <- "RT @HenryHill is the best dog ever #InstaHen"
twit_pre(text_str)

}
