% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vader}
\alias{vader}
\title{Vader Lexicon}
\format{
A \code{\link[base]{data.frame}} with 7,495 rows and 2 columns
}
\source{
\link{https://github.com/cjhutto/vaderSentiment}
}
\description{
A dataframe containing the preprocessed and deduplicated Vader lexicon (Hutto & Gilbert 2014). The original dataset was preprocessed using \pkg{WoCSentiment} function \code{cleanse_lex}.
}
\details{
\itemize{
\item word. An n-gram representing a term or phrase
\item value. The sentiment intensity score
}
}
\section{License}{
 The MIT License (MIT).
}

\section{Copyright}{
 (c) 2016 C.J. Hutto. 

Permission is hereby granted, free of charge, to any person obtaining a copy
of this software and associated documentation files (the "Software"), to deal
in the Software without restriction, including without limitation the rights
to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
copies of the Software, and to permit persons to whom the Software is
furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all
copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
SOFTWARE.
}

\references{
Hutto, C.J. and Gilbert, E.E. (2014). VADER: A Parsimonious Rule-based Model
for Sentiment Analysis of Social Media Text. Eighth International Conference
on Weblogs and Social Media (ICWSM-14). Ann Arbor, MI, June 2014.
}
