% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wkwsci}
\alias{wkwsci}
\title{WKWSCI Lexicon}
\format{
A \code{\link[base]{data.frame}} with 10,181 rows and 2 columns
}
\source{
\link{https://researchdata.ntu.edu.sg/dataset.xhtml?persistentId=doi:10.21979/N9/DWWEBV}
}
\description{
The preprocessed and deduplicated SO-CAL lexicon (Khoo & Johnkhan 2017). The original dataset was preprocessed using \pkg{WoCSentiment} function \code{cleanse_lex}.
}
\details{
\itemize{
\item word. An n-gram representing a term or phrase
\item value. The sentiment intensity score

}
}
\section{License}{
 This work is licensed under a Creative Commons 
Attribution-NonCommercial-ShareAlike 4.0 International License.
https://creativecommons.org/licenses/by-nc-sa/4.0/
}

\references{
Khoo, C. S., & Johnkhan, S. B. (2017). Lexicon-based sentiment analysis: Comparative evaluation of six sentiment lexicons. Journal of Information Science, pp. 1–21.
}
