% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WoC_Crowd}
\alias{WoC_Crowd}
\title{WoC Lexicons}
\format{
A \code{\link[base]{vector}} with 13 character-valued lexicon object names
}
\description{
A vector including 13 character strings representing the names of the 13 lexicons included in the default WoC crowd in the \code{score_woc} function.
}
